Fixed2.0

+++ contracts/ERC20/USDT.sol
@@ -5,7 +5,7 @@ pragma solidity 0.8.20;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract USDT is ERC20 {
-    constructor(uint amount) ERC20("USDT", "USDT") {
+    constructor(uint256 amount) ERC20("USDT", "USDT") {
         _mint(msg.sender, amount);
     }
 }
